+++ packages/customize-widgets/src/components/sidebar-block-editor/sidebar-editor-provider.js
@@ -1,32 +1,19 @@
 /**
  * WordPress dependencies
  */
-import { useMemo } from '@wordpress/element';
-import {
-	BlockEditorProvider,
-	BlockEditorKeyboardShortcuts,
-} from '@wordpress/block-editor';
+import { BlockEditorProvider } from '@wordpress/block-editor';
 
 /**
  * Internal dependencies
  */
 import useSidebarBlockEditor from './use-sidebar-block-editor';
-import Header from '../header';
-import useInserter from '../inserter/use-inserter';
 
 export default function SidebarEditorProvider( {
 	sidebar,
-	inserter,
+	settings,
 	children,
 } ) {
 	const [ blocks, onInput, onChange ] = useSidebarBlockEditor( sidebar );
-	const [ isInserterOpened, setIsInserterOpened ] = useInserter( inserter );
-	const settings = useMemo(
-		() => ( {
-			__experimentalSetIsInserterOpened: setIsInserterOpened,
-		} ),
-		[]
-	);
 
 	return (
 		<BlockEditorProvider
@@ -36,14 +23,6 @@ export default function SidebarEditorProvider( {
 			settings={ settings }
 			useSubRegistry={ false }
 		>
-			<BlockEditorKeyboardShortcuts />
-
-			<Header
-				inserter={ inserter }
-				isInserterOpened={ isInserterOpened }
-				setIsInserterOpened={ setIsInserterOpened }
-			/>
-
 			{ children }
 		</BlockEditorProvider>
 	);
